reverted changes to gitignore, moved and merged TestZkMaintenanceUtils
